+++ packages/components/src/Form/Inputs/InputSearch/InputSearch.tsx
@@ -90,6 +90,8 @@ const InputSearchLayout = forwardRef(
     {
       options,
       disabled,
+      autoFocus,
+      'data-autofocus': dataAutofocus,
       isClearable = true,
       placeholder,
       name,
@@ -169,6 +171,8 @@ const InputSearchLayout = forwardRef(
           iconBefore={hideSearchIcon ? undefined : 'Search'}
           iconBeforeTitle={hideSearchIcon ? undefined : 'Search'}
           disabled={disabled}
+          autoFocus={autoFocus}
+          data-autofocus={dataAutofocus}
           placeholder={placeholder}
           name={name}
           validationType={props.validationType}