+++ packages/components/src/Form/Inputs/InputTimeSelect/InputTimeSelect.tsx
@@ -86,6 +86,10 @@ export interface InputTimeSelectProps
   onChange?: (val?: string) => void
   validationType?: ValidationType
   onValidationFail?: (value: string) => void
+  /**
+   * Sets initial focus when inside a Dialog or Popover
+   */
+  'data-autofocus'?: string
 }
 
 // if format is `12h`, repeat hours 1-12 twice
@@ -276,6 +280,8 @@ const InputTimeSelectLayout = forwardRef(
       defaultValue,
       validationType,
       disabled,
+      autoFocus,
+      'data-autofocus': dataAutofocus,
       id,
       ...props
     }: InputTimeSelectProps,
@@ -383,6 +389,8 @@ const InputTimeSelectLayout = forwardRef(
           autoComplete={false}
           validationType={validationType}
           disabled={disabled}
+          autoFocus={autoFocus}
+          data-autofocus={dataAutofocus}
           id={id}
           {...ariaProps}
         />