+++ packages/components/src/Form/Inputs/Select/Select.tsx
@@ -74,6 +74,10 @@ export interface SelectBaseProps
    * defaults to false for <100 options, true for >=100 options
    */
   windowedOptions?: boolean
+  /**
+   * Sets initial focus when inside a Dialog or Popover
+   */
+  'data-autofocus'?: string
 }
 
 export interface SelectProps
@@ -103,6 +107,8 @@ const SelectComponent = forwardRef(
     {
       options,
       disabled,
+      autoFocus,
+      'data-autofocus': dataAutofocus,
       isFilterable,
       isClearable,
       placeholder,
@@ -162,6 +168,8 @@ const SelectComponent = forwardRef(
           {...ariaProps}
           before={<SelectInputIcon options={options} />}
           disabled={disabled}
+          autoFocus={autoFocus}
+          data-autofocus={dataAutofocus}
           placeholder={placeholder}
           name={name}
           validationType={props.validationType}